requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.4
backcall==0.1.0
bleach==3.1.5
branca==0.4.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
djangorestframework==3.11.1
entrypoints==0.3
folium==0.11.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
queuelib==1.5.0
requests==2.23.0
Scrapy==2.1.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
zope.interface==5.1.0
